# perf: don't create duplicate function/object templates for ops